requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
audioop-lts==0.2.2
audioread==3.0.1
av==15.1.0
azure-cognitiveservices-speech==1.46.0
azure-core==1.35.1
beautifulsoup4==4.13.5
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
clr==1.0.3
clr_loader==0.2.7.post0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
comtypes==1.4.12
ctranslate2==4.6.0
dataclasses-json==0.6.7
debugpy==1.8.16
decorator==5.2.1
deepgram-sdk==3.4.0
distro==1.9.0
dnspython==2.8.0
docutils==0.22
elevenlabs==0.2.26
pygame==2.6.1
emoji==2.14.1
enum34==1.1.10
eventlet==0.33.3
executing==2.2.1
faster-whisper==1.1.1
filelock==3.19.1
filetype==1.2.0
Flask==2.3.3
Flask-SocketIO==5.3.6
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.9.0
fuzzywuzzy==0.18.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.31.1
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
halo==0.0.31
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
HumanCursor==1.1.5
humanfriendly==10.0
id==1.5.0
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
keyboard==0.13.5
keyring==25.6.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-google-genai==2.1.10
langchain-groq==0.3.8
langchain-ollama==0.3.8
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.27
lazy_loader==0.4
Levenshtein==0.27.1
librosa==0.11.0
live-inspect==0.1.1
llvmlite==0.44.0
log-symbols==0.0.14
markdown-it-py==4.0.0
markdownify==1.2.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.8.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nh3==0.3.0
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
ollama==0.5.3
onnxruntime==1.22.1
openai==1.107.2
openwakeword==0.6.0
orjson==3.11.3
ormsgpack==1.10.0
outcome==1.3.0.post0
packaging==25.0
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
pure_eval==0.2.3
pvporcupine==1.9.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyperclip==1.9.0
pypiwin32==223
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.7.1
python-Levenshtein==0.27.1
python-socketio==5.9.0
pythonnet==3.0.5
pytweening==1.2.0
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==27.1.0
RapidFuzz==3.14.1
ratelimit==2.2.1
readme_renderer==44.0
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.1.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.2
selenium==4.35.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
spinners==0.0.24
SQLAlchemy==2.0.43
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
stanza==1.10.1
stream2sentence==0.3.1
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.8.0
torchaudio==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
twine==6.2.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uiautomation==2.0.29
urllib3==2.5.0
uv==0.8.17
wcwidth==0.2.13
webrtcvad-wheels==2.0.14
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
windows-use==0.6.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0